Link to 32-bit downloads for Windows instead of 64-bit
We still have Windows users on tablet devices that are apparently
32-bit. DCSS doesn't really need to be a 64-bit application, and having
builds for both architectures is unwieldy so long as we don't have
automated Windows builds nor a way to organize files on the website
(other than simply putting them in the releases folder). Untill both
issues can be addressed, we'll simply do 32-bit builds and link to
those, so we can support the maximum number of platforms. 32-bit builds
will work fine on 64-bit windows systems, after all.