Replace live games system
The old live games system used a combination of 'dgl-status' polling for
a master games list and Sequell HTTP API milestone queries for detailed
game information. Replace this system with a single API call to the
crawl live games API I'm hosting.

The old system had a few problems:
* Not every server had a dgl-status, so their games would never be
  featured on the front page (notably CJR, but other servers have had
  their dgl-status broken for long periods of time).
* Reliance on a local cronjob on the website's webserver (which makes
  the website unsuitable for hosting on a CDN).
* Reliance on sequell milestone API, which doesn't get fixed quickly
  when there are problems.

The live games API I've written is pretty new, and still not perfect.
However the API centralises data that was previously synthesised from
two sources and thus greatly simplifies the code on the website's end.

Having said that, a the Javascript I've written in dcss.js is, on the
whole, terrible quality. Improvements would be most welcome...