Update Download page for 0.25
I've linked directly to the github release packages, since these are
built automatically and hosted on github. We might want to break up Past
Releases into those we hosted on CDO and those we have on github, which
will presumably be from now on.

Also link the listed release version to the github release, since that
has a nice rendition of the changelog.